+++ packages/styles/src/packages/atomic/toggle/rarui-toggle.types.ts
@@ -1,0 +1,4 @@
+import { RecipeVariants } from "@vanilla-extract/recipes";
+import { styles } from "./rarui-toggle.css";
+
+export type ToggleVariants = RecipeVariants<typeof styles.container__slider>;