+++ packages/styles/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 RarUI Styles deprive all styles needed to build components.
 
+## 2024-04-18 `1.11.0`
+
+#### 🎉 New features
+
+- Added new style pack for `toggle` component. ([#25](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/25) by [@junior](https://git.rarolabs.com.br/junior))
+
 ## 2024-04-18 `1.10.0`
 
 #### 🎉 New features